requirements.txt
sentence-transformers
numpy==1.26.4
python-dotenv
pinecone-client
streamlit
google-cloud-language
PyMuPDF 
python-docx